workers/crawler/requirements.txt
requests
beautifulsoup4
pdfminer.six
tiktoken